backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.21
pydantic>=2.4.2
psycopg2-binary>=2.9.7
python-dotenv>=1.0.0
alembic>=1.12.0
websockets>=11.0.3
pytest>=7.4.2
httpx>=0.25.0
openai>=1.3.0
langchain>=0.0.335
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
